disk_file_c: Fix a warning
warning C4244: 'argument' : conversion from 'uint64_t' to 'long', possible loss of data

+++ disk_file_c.cpp
@@ -49,7 +49,7 @@ public:
 
 	void seek(uint64_t pos)
 	{
-		fseek(archive, pos * sizeof(NODE), SEEK_SET);
+		fseek(archive, (long)(pos * sizeof(NODE)), SEEK_SET);
 	}
 
 	void close()
